# Patch Notes v33.0