… bundleV2/f029e3f32e288984.mod_bundle
… scripts/mods/TourneyBalance/TourneyBalance_localization.lua
… scripts/mods/TourneyBalance/changes/career_changes.lua
… scripts/mods/TourneyBalance/logging_and_qol/mod_checker.lua
… scripts/mods/TourneyBalance/logging_and_qol/performance_logging.lua