- Adjusted Maximum Values for Text position to prevent it from being pushed off screen
- Added functionality to hide the unnaproved mods showing on screen (only the status of the mods will still show)
- Grail Knight Potion Quest values have been adjusted to scale in instances of 50 kills instead of 25. (Results in Cataclysm 1 = 350, Cataclysm 3 = 500)
- Virtue of Stoicism reduced from 50% to 25%
- Vengeance buffed to Time between stacks, 3 -> 2 seconds, AS per stack, 6% ->4.8%, Duration 12 -> 15 seconds
- Under Pressure nerfed. Damage to high heat lowered significantly, Attack speed at high heat lowered significantly
- Waystalker passive regen heal cap is now 66.6% instead of 100%, still stays 100% when using Isha's Embrace
- Salvaged ammunition also procs on Specials now
- THP on crit reverted to previous Tourney Balance version before the fatshark rework of the thp talent row
- Bolt: Increased Bolt projectile Lifetime (No dissappearing long range bolts anymore)
- Sienna Dagger buffed, mostly affects BW and Famished Flames as the PA has a dot now

- Fixed Mod Checker System to not spam the same lines
- Fixed Mod Checker System to communicate properly between players
- Adjusted Status line to be more clear
- Added option to remove list of unapproved mod

- Adjusted Performance Logging to account for changes in Linesman
- Adjusted Performance Logging output slightly to be a bit clearer and easier to use (Google sheet has template v3 now)
- Fixed Performance logging not working properly when restarting